Renamed LCGrid as it is no longer a linked cell grid.
The class is responsible for the neighborhood search and we are doing
this with a regular 2D grid at the moment.
