+++ libcore/CMakeLists.txt
@@ -48,8 +48,8 @@ set(source_files
     src/math/Mathematics.cpp
     src/math/OperationalModel.cpp
     src/math/VelocityModel.cpp
-    src/mpi/LCGrid.cpp
-    src/mpi/Grid2D.cpp
+    src/neighborhood/NeighborhoodSearch.cpp
+    src/neighborhood/Grid2D.cpp
     src/pedestrian/AgentsParameters.cpp
     src/pedestrian/AgentsQueue.cpp
     src/pedestrian/AgentsSource.cpp
@@ -136,8 +136,8 @@ set(header_files
     src/math/Mathematics.h
     src/math/OperationalModel.h
     src/math/VelocityModel.h
-    src/mpi/LCGrid.h
-    src/mpi/Grid2D.h
+    src/neighborhood/NeighborhoodSearch.h
+    src/neighborhood/Grid2D.h
     src/pedestrian/AgentsParameters.h
     src/pedestrian/AgentsQueue.h
     src/pedestrian/AgentsSource.h
