+++ libcore/src/neighborhood/Grid2D.cpp
